moving sri digest out to sourceable script
