+++ scripts/integrity.js
@@ -1,9 +1,8 @@
 #!/usr/bin/env node
 var path = require('path');
 var yaml = require('js-yaml');
-var exec = require('child_process').execSync;
 var fs   = require('fs');
-var sri  = require('sri-toolbox');
+var sri  = require(path.join(__dirname, 'sri'));
 
 var basedir    = path.join(__dirname, '..');
 var configFile = path.join(basedir, 'config', '_config.yml');
@@ -16,13 +15,6 @@ fs.createReadStream(configFile)
 function buildPath(d) {
     return path.join(basedir, 'public',
         d.replace('https://maxcdn.bootstrapcdn.com/',''));
-}
-
-function digest(file) {
-    return 'sha256-' +
-        exec('cat '  + file + ' | openssl dgst -sha256 -binary | openssl enc -base64 -A').toString()  +
-        ' sha512-'   +
-        exec('cat '  + file + ' | openssl dgst -sha512 -binary | openssl enc -base64 -A').toString();
 }
 
 function exists(file) {
@@ -41,7 +33,7 @@ for (var i = 0; i < config.bootswatch.themes.length; i++) {
                           .replace('SWATCH_NAME',    theme.name);
 
     if (exists(file)) { // always regen
-        config.bootswatch.themes[i].sri = digest(file);
+        config.bootswatch.themes[i].sri = sri.digest(file);
     }
 }
 
@@ -51,7 +43,7 @@ for (var i = 0; i < config.bootlint.length; i++) {
     var file     = buildPath(bootlint.javascript);
 
     if (config.bootlint[i].javascript_sri === undefined && exists(file)) {
-        config.bootlint[i].javascript_sri = digest(file);
+        config.bootlint[i].javascript_sri = sri.digest(file);
     }
 }
 
@@ -61,12 +53,12 @@ for (var i = 0; i < config.bootlint.length; i++) {
         var bootstrap  = config[key][i];
         var javascript = buildPath(bootstrap.javascript);
         if (config[key][i].javascript_sri === undefined && exists(javascript)) {
-            config[key][i].javascript_sri = digest(javascript);
+            config[key][i].javascript_sri = sri.digest(javascript);
         }
 
         var stylesheet = buildPath(bootstrap.stylesheet);
         if (config[key][i].stylesheet_sri === undefined && exists(stylesheet)) {
-            config[key][i].stylesheet_sri = digest(stylesheet);
+            config[key][i].stylesheet_sri = sri.digest(stylesheet);
         }
     }
 });
@@ -76,7 +68,7 @@ for (var i = 0; i < config.fontawesome.length; i++) {
     var stylesheet = buildPath(config.fontawesome[i].stylesheet);
 
     if (config.fontawesome[i].stylesheet_sri === undefined && exists(stylesheet)) {
-        config.fontawesome[i].stylesheet_sri = digest(stylesheet);
+        config.fontawesome[i].stylesheet_sri = sri.digest(stylesheet);
     }
 }
